# soroban-rpc: Cache all ledger entries queried from DB in read transaction