--- a/cmd/soroban-rpc/internal/methods/simulate_transaction.go
+++ b/cmd/soroban-rpc/internal/methods/simulate_transaction.go
@@ -84,7 +84,7 @@ func NewSimulateTransactionHandler(logger *log.Entry, ledgerEntryReader db.Ledge
 			}
 		}
 
-		readTx, err := ledgerEntryReader.NewTx(ctx)
+		readTx, err := ledgerEntryReader.NewCachedTx(ctx)
 		if err != nil {
 			return SimulateTransactionResponse{
 				Error: "Cannot create read transaction",