@2opremio 2023-08-08
> this will not cache lookups for entries which don't exist.

Uhm, good point, maybe I should fix this.