@2opremio 2023-08-08
I fixed it and now the cache performs even better:

```
goos: darwin
goarch: arm64
pkg: github.com/stellar/soroban-tools/cmd/soroban-rpc/internal/preflight
BenchmarkGetPreflight
BenchmarkGetPreflight/In-memory_storage
BenchmarkGetPreflight/In-memory_storage-12         	    6151	    194964 ns/op
BenchmarkGetPreflight/DB_storage
BenchmarkGetPreflight/DB_storage-12                	    4008	    274884 ns/op
PASS
```

we are _only_ 0.7 ms (40% more) away from the in-memory storage.